Refactor DatabaseAccess and add JabMapHTTPServer
