+++ src/main/java/org/jabref/gui/JabRefMain.java
@@ -25,6 +25,7 @@ import org.jabref.logic.util.OS;
 import org.jabref.migrations.PreferencesMigrations;
 import org.jabref.model.database.BibDatabaseMode;
 import org.jabref.preferences.JabRefPreferences;
+import org.jabref.rest.resources.JabMapHTTPServer;
 import org.jabref.rest.resources.RootResource;
 
 import org.apache.commons.cli.ParseException;
@@ -76,7 +77,7 @@ public class JabRefMain extends Application {
                     return;
                 }
 
-                startHttpEndPoint();
+                JabMapHTTPServer.startHttpEndPoint();
 
                 // TODO: How and when to shut down the server
                 // Waits until server is finished.
@@ -96,40 +97,6 @@ public class JabRefMain extends Application {
             LOGGER.error("Unexpected exception", ex);
             Platform.exit();
         }
-    }
-
-    private void startHttpEndPoint() {
-        Server server = this.createHttpServer();
-
-        try {
-            // Starts server to http://localhost:9898/
-            // The current implementation serves libraries/current/entries
-            server.start();
-        } catch (Exception e) {
-            LOGGER.error("Problem starting HTTP Server", e);
-        }
-    }
-
-    private Server createHttpServer() {
-        ServletContextHandler context = new ServletContextHandler(ServletContextHandler.SESSIONS);
-        context.setContextPath("/");
-
-        Server server = new Server(9898);
-        server.setHandler(context);
-        this.addServlet(context);
-        return server;
-    }
-
-    private void addServlet(ServletContextHandler context) {
-        // this mirrors a webapp/WEB-INF/web.xml
-        ServletHolder jerseyServlet = context.addServlet(
-                org.glassfish.jersey.servlet.ServletContainer.class, "/*");
-        jerseyServlet.setInitOrder(0);
-
-        // Tells the Jersey Servlet which REST service/class to load.
-        jerseyServlet.setInitParameter(
-                "jersey.config.server.provider.classnames",
-                RootResource.class.getCanonicalName());
     }
 
     @Override
