+++ src/test/java/org/jabref/rest/resources/RootResourceTest.java
@@ -46,43 +46,9 @@ class RootResourceTest {
 
     private static Client client;
 
-    private static void startHttpEndPoint() {
-        Server server = createHttpServer();
-
-        try {
-            // Starts server to http://localhost:9898/
-            // The current implementation serves libraries/current/entries
-            server.start();
-        } catch (Exception e) {
-//            LOGGER.error("Problem starting HTTP Server", e);
-        }
-    }
-
-    private static Server createHttpServer() {
-        ServletContextHandler context = new ServletContextHandler(ServletContextHandler.SESSIONS);
-        context.setContextPath("/");
-
-        Server server = new Server(9898);
-        server.setHandler(context);
-        addServlet(context);
-        return server;
-    }
-
-    private static void addServlet(ServletContextHandler context) {
-        // this mirrors a webapp/WEB-INF/web.xml
-        ServletHolder jerseyServlet = context.addServlet(
-                org.glassfish.jersey.servlet.ServletContainer.class, "/*");
-        jerseyServlet.setInitOrder(0);
-
-        // Tells the Jersey Servlet which REST service/class to load.
-        jerseyServlet.setInitParameter(
-                "jersey.config.server.provider.classnames",
-                RootResource.class.getCanonicalName());
-    }
-
     @BeforeAll
     static void setUp() {
-        startHttpEndPoint();
+        JabMapHTTPServer.startHttpEndPoint();
         client = ClientBuilder.newClient();
     }
 
@@ -163,6 +129,7 @@ class RootResourceTest {
             assertEquals(expectedBibEntries.size(), actualBibEntries.size());
             for (BibEntry bibEntry : expectedBibEntries) {
                 assertTrue(actualBibEntries.contains(bibEntry));
+                actualBibEntries.remove(bibEntry);
             }
         } finally {
             // Close the Response object.
